refactor #7 : 반환 형태 수정
